requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
google-generativeai==0.3.2
